+++ spec/loan_creator/bullet_spec.rb
@@ -13,10 +13,6 @@ describe LoanCreator::Bullet do
       context "for scenario #{loan_type}_#{scenario.join('_')}" do
         include_examples('valid lender timetable', loan_type, scenario)
       end
-
-      context "for scenario realistic_#{loan_type}_#{scenario.join('_')} with realistic durations" do
-        include_examples('valid lender timetable', loan_type, scenario, {}, true)
-      end
     end
 
     context 'given a scenario with initial_values' do
@@ -31,7 +27,16 @@ describe LoanCreator::Bullet do
       scenario = ['year', '55000', '10', '2018-01-01', '3', '0']
 
       include_examples('valid lender timetable', loan_type, scenario, initial_values)
-      include_examples('valid lender timetable', loan_type, scenario, initial_values, true)
+    end
+
+    context 'given scenarios with realistic durations' do
+      scenarios = [
+        ['semester', '100000', '12', '2022-04-15', '3', '0'],
+        ['semester', '100000', '12', '2021-04-15', '3', '0']
+      ]
+      scenarios.each do |scenario|
+        include_examples('valid lender timetable', loan_type, scenario, nil, true)
+      end
     end
 
     context 'given a scenario with term dates' do