+++ spec/loan_creator/in_fine_spec.rb
@@ -16,10 +16,6 @@ describe LoanCreator::InFine do
       context "for scenario #{loan_type}_#{scenario.join('_')}" do
         include_examples('valid lender timetable', loan_type, scenario)
       end
-
-      context "for scenario realistic_#{loan_type}_#{scenario.join('_')} with realistic durations" do
-        include_examples('valid lender timetable', loan_type, scenario, {}, true)
-      end
     end
 
     context 'given a scenario with initial_values' do
@@ -34,7 +30,16 @@ describe LoanCreator::InFine do
       scenario = ['year', '55000', '10', '2018-01-01', '3', '0']
 
       include_examples('valid lender timetable', loan_type, scenario, initial_values)
-      include_examples('valid lender timetable', loan_type, scenario, initial_values, true)
+    end
+
+    context 'given scenarios with realistic durations' do
+      scenarios = [
+        ['month', '100000', '12', '2021-12-15', '18', '0'],
+        ['quarter', '100000', '12', '2021-01-15', '12', '0']
+      ]
+      scenarios.each do |scenario|
+        include_examples('valid lender timetable', loan_type, scenario, nil, true)
+      end
     end
 
     context 'given a scenario with term dates' do